# portal/api/cron/v1/cron_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.0
// source: cron/v1/cron.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationCronCreateCronTask = "/cron.v1.Cron/CreateCronTask"
const OperationCronDeleteCronTask = "/cron.v1.Cron/DeleteCronTask"
const OperationCronGetCronTask = "/cron.v1.Cron/GetCronTask"
const OperationCronListCronRuns = "/cron.v1.Cron/ListCronRuns"
const OperationCronListCronTasks = "/cron.v1.Cron/ListCronTasks"
const OperationCronRunCronTask = "/cron.v1.Cron/RunCronTask"
const OperationCronUpdateCronTask = "/cron.v1.Cron/UpdateCronTask"

type CronHTTPServer interface {
	// CreateCronTask 创建定时任务
	CreateCronTask(context.Context, *CreateCronTaskRequest) (*CronTaskReply, error)
	// DeleteCronTask 删除任务
	DeleteCronTask(context.Context, *DeleteCronTaskRequest) (*DeleteCronTaskReply, error)
	// GetCronTask 获取单个任务
	GetCronTask(context.Context, *GetCronTaskRequest) (*CronTaskReply, error)
	// ListCronRuns 执行历史
	ListCronRuns(context.Context, *ListCronRunsRequest) (*ListCronRunsReply, error)
	// ListCronTasks 任务列表（分页、筛选）
	ListCronTasks(context.Context, *ListCronTasksRequest) (*ListCronTasksReply, error)
	// RunCronTask 立即执行
	RunCronTask(context.Context, *RunCronTaskRequest) (*RunCronTaskReply, error)
	// UpdateCronTask 更新任务
	UpdateCronTask(context.Context, *UpdateCronTaskRequest) (*CronTaskReply, error)
}

func RegisterCronHTTPServer(s *http.Server, srv CronHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/cron/tasks", _Cron_CreateCronTask0_HTTP_Handler(srv))
	r.GET("/api/v1/cron/tasks", _Cron_ListCronTasks0_HTTP_Handler(srv))
	r.GET("/api/v1/cron/tasks/{id}", _Cron_GetCronTask0_HTTP_Handler(srv))
	r.PUT("/api/v1/cron/tasks/{id}", _Cron_UpdateCronTask0_HTTP_Handler(srv))
	r.DELETE("/api/v1/cron/tasks/{id}", _Cron_DeleteCronTask0_HTTP_Handler(srv))
	r.POST("/api/v1/cron/tasks/{id}/run", _Cron_RunCronTask0_HTTP_Handler(srv))
	r.GET("/api/v1/cron/tasks/{id}/runs", _Cron_ListCronRuns0_HTTP_Handler(srv))
}

func _Cron_CreateCronTask0_HTTP_Handler(srv CronHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateCronTaskRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCronCreateCronTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateCronTask(ctx, req.(*CreateCronTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CronTaskReply)
		return ctx.Result(200, reply)
	}
}

func _Cron_ListCronTasks0_HTTP_Handler(srv CronHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListCronTasksRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCronListCronTasks)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListCronTasks(ctx, req.(*ListCronTasksRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListCronTasksReply)
		return ctx.Result(200, reply)
	}
}

func _Cron_GetCronTask0_HTTP_Handler(srv CronHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCronTaskRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCronGetCronTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCronTask(ctx, req.(*GetCronTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CronTaskReply)
		return ctx.Result(200, reply)
	}
}

func _Cron_UpdateCronTask0_HTTP_Handler(srv CronHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateCronTaskRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCronUpdateCronTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateCronTask(ctx, req.(*UpdateCronTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CronTaskReply)
		return ctx.Result(200, reply)
	}
}

func _Cron_DeleteCronTask0_HTTP_Handler(srv CronHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteCronTaskRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCronDeleteCronTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteCronTask(ctx, req.(*DeleteCronTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteCronTaskReply)
		return ctx.Result(200, reply)
	}
}

func _Cron_RunCronTask0_HTTP_Handler(srv CronHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RunCronTaskRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCronRunCronTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RunCronTask(ctx, req.(*RunCronTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RunCronTaskReply)
		return ctx.Result(200, reply)
	}
}

func _Cron_ListCronRuns0_HTTP_Handler(srv CronHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListCronRunsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCronListCronRuns)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListCronRuns(ctx, req.(*ListCronRunsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListCronRunsReply)
		return ctx.Result(200, reply)
	}
}

type CronHTTPClient interface {
	// CreateCronTask 创建定时任务
	CreateCronTask(ctx context.Context, req *CreateCronTaskRequest, opts ...http.CallOption) (rsp *CronTaskReply, err error)
	// DeleteCronTask 删除任务
	DeleteCronTask(ctx context.Context, req *DeleteCronTaskRequest, opts ...http.CallOption) (rsp *DeleteCronTaskReply, err error)
	// GetCronTask 获取单个任务
	GetCronTask(ctx context.Context, req *GetCronTaskRequest, opts ...http.CallOption) (rsp *CronTaskReply, err error)
	// ListCronRuns 执行历史
	ListCronRuns(ctx context.Context, req *ListCronRunsRequest, opts ...http.CallOption) (rsp *ListCronRunsReply, err error)
	// ListCronTasks 任务列表（分页、筛选）
	ListCronTasks(ctx context.Context, req *ListCronTasksRequest, opts ...http.CallOption) (rsp *ListCronTasksReply, err error)
	// RunCronTask 立即执行
	RunCronTask(ctx context.Context, req *RunCronTaskRequest, opts ...http.CallOption) (rsp *RunCronTaskReply, err error)
	// UpdateCronTask 更新任务
	UpdateCronTask(ctx context.Context, req *UpdateCronTaskRequest, opts ...http.CallOption) (rsp *CronTaskReply, err error)
}

type CronHTTPClientImpl struct {
	cc *http.Client
}

func NewCronHTTPClient(client *http.Client) CronHTTPClient {
	return &CronHTTPClientImpl{client}
}

// CreateCronTask 创建定时任务
func (c *CronHTTPClientImpl) CreateCronTask(ctx context.Context, in *CreateCronTaskRequest, opts ...http.CallOption) (*CronTaskReply, error) {
	var out CronTaskReply
	pattern := "/api/v1/cron/tasks"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationCronCreateCronTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteCronTask 删除任务
func (c *CronHTTPClientImpl) DeleteCronTask(ctx context.Context, in *DeleteCronTaskRequest, opts ...http.CallOption) (*DeleteCronTaskReply, error) {
	var out DeleteCronTaskReply
	pattern := "/api/v1/cron/tasks/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCronDeleteCronTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetCronTask 获取单个任务
func (c *CronHTTPClientImpl) GetCronTask(ctx context.Context, in *GetCronTaskRequest, opts ...http.CallOption) (*CronTaskReply, error) {
	var out CronTaskReply
	pattern := "/api/v1/cron/tasks/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCronGetCronTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListCronRuns 执行历史
func (c *CronHTTPClientImpl) ListCronRuns(ctx context.Context, in *ListCronRunsRequest, opts ...http.CallOption) (*ListCronRunsReply, error) {
	var out ListCronRunsReply
	pattern := "/api/v1/cron/tasks/{id}/runs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCronListCronRuns))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListCronTasks 任务列表（分页、筛选）
func (c *CronHTTPClientImpl) ListCronTasks(ctx context.Context, in *ListCronTasksRequest, opts ...http.CallOption) (*ListCronTasksReply, error) {
	var out ListCronTasksReply
	pattern := "/api/v1/cron/tasks"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCronListCronTasks))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RunCronTask 立即执行
func (c *CronHTTPClientImpl) RunCronTask(ctx context.Context, in *RunCronTaskRequest, opts ...http.CallOption) (*RunCronTaskReply, error) {
	var out RunCronTaskReply
	pattern := "/api/v1/cron/tasks/{id}/run"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCronRunCronTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateCronTask 更新任务
func (c *CronHTTPClientImpl) UpdateCronTask(ctx context.Context, in *UpdateCronTaskRequest, opts ...http.CallOption) (*CronTaskReply, error) {
	var out CronTaskReply
	pattern := "/api/v1/cron/tasks/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationCronUpdateCronTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}