# portal/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/cron/tasks:
        get:
            tags:
                - Cron
            description: 任务列表（分页、筛选）
            operationId: Cron_ListCronTasks
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: agentId
                  in: query
                  schema:
                    type: string
                - name: enabled
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/cron.v1.ListCronTasksReply'
        post:
            tags:
                - Cron
            description: 创建定时任务
            operationId: Cron_CreateCronTask
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/cron.v1.CreateCronTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/cron.v1.CronTaskReply'
    /api/v1/cron/tasks/{id}:
        get:
            tags:
                - Cron
            description: 获取单个任务
            operationId: Cron_GetCronTask
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/cron.v1.CronTaskReply'
        put:
            tags:
                - Cron
            description: 更新任务
            operationId: Cron_UpdateCronTask
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/cron.v1.UpdateCronTaskRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/cron.v1.CronTaskReply'
        delete:
            tags:
                - Cron
            description: 删除任务
            operationId: Cron_DeleteCronTask
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/cron.v1.DeleteCronTaskReply'
    /api/v1/cron/tasks/{id}/run:
        post:
            tags:
                - Cron
            description: 立即执行
            operationId: Cron_RunCronTask
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/cron.v1.RunCronTaskReply'
    /api/v1/cron/tasks/{id}/runs:
        get:
            tags:
                - Cron
            description: 执行历史
            operationId: Cron_ListCronRuns
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/cron.v1.ListCronRunsReply'
...
components:
    schemas:
...
        api.common.BaseResponse:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                reason:
                    type: string
            description: 统一响应基类，所有 Response/Reply 需嵌入
...
        cron.v1.CreateCronTaskRequest:
            type: object
            properties:
                name:
                    type: string
                agentId:
                    type: string
                scheduleKind:
                    type: string
                scheduleExpr:
                    type: string
                timezone:
                    type: string
                staggerSec:
                    type: integer
                    format: int32
                payloadKind:
                    type: string
                payloadContent:
                    type: string
                timeoutSec:
                    type: integer
                    format: int32
                retryCount:
                    type: integer
                    format: int32
                retryIntervalSec:
                    type: integer
                    format: int32
                deliveryMode:
                    type: string
                deliveryWebhookUrl:
                    type: string
                deliverySecret:
                    type: string
                deliveryBestEffort:
                    type: boolean
                deliverySessionId:
                    type: string
                deliveryChannelId:
                    type: string
                enabled:
                    type: boolean
        cron.v1.CronRunReply:
            type: object
            properties:
                id:
                    type: string
                taskId:
                    type: string
                triggeredAt:
                    type: string
                startedAt:
                    type: string
                finishedAt:
                    type: string
                status:
                    type: string
                outputSummary:
                    type: string
                error:
                    type: string
                deliveryOk:
                    type: boolean
        cron.v1.CronTaskReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
                id:
                    type: string
                name:
                    type: string
                agentId:
                    type: string
                scheduleKind:
                    type: string
                scheduleExpr:
                    type: string
                timezone:
                    type: string
                staggerSec:
                    type: integer
                    format: int32
                payloadKind:
                    type: string
                payloadContent:
                    type: string
                timeoutSec:
                    type: integer
                    format: int32
                retryCount:
                    type: integer
                    format: int32
                retryIntervalSec:
                    type: integer
                    format: int32
                deliveryMode:
                    type: string
                deliveryWebhookUrl:
                    type: string
                deliverySessionId:
                    type: string
                deliveryChannelId:
                    type: string
                enabled:
                    type: boolean
                nextRunAt:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
        cron.v1.DeleteCronTaskReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
        cron.v1.ListCronRunsReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/cron.v1.CronRunReply'
                total:
                    type: integer
                    format: int32
        cron.v1.ListCronTasksReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/cron.v1.CronTaskReply'
                total:
                    type: integer
                    format: int32
        cron.v1.RunCronTaskReply:
            type: object
            properties:
                ret:
                    $ref: '#/components/schemas/api.common.BaseResponse'
        cron.v1.UpdateCronTaskRequest:
            type: object
            properties:
                id:
                    type: string
                updates:
                    type: object
...